+++ desktop/src/renderer/lib/constants.ts
@@ -1,8 +1,8 @@
 export const APP_NAME = "Brewwery";
-export const APP_VERSION = "0.9.5";
+export const APP_VERSION = "0.9.6";
 export const APP_CHANNEL = "Release Candidate";
 export const GITHUB_URL = "https://github.com/brewwery/brewwery";
 export const WEBSITE_URL = "https://www.brewwery.com";
 export const ISSUE_URL = "https://github.com/brewwery/brewwery/issues";
-export const RELEASE_NOTES_URL = "https://github.com/brewwery/brewwery/releases/tag/v0.9.5";
+export const RELEASE_NOTES_URL = "https://github.com/brewwery/brewwery/releases/tag/v0.9.6";
 export const HOME_BREW_URL = "https://brew.sh";